% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add}
\alias{add}
\title{Add x and y}
\usage{
add(x, y)
}
\arguments{
\item{x}{first argument of the sum}

\item{y}{second argument of the sum}
}
\value{
x + y
}
\description{
This is a helper function for use in Reduce() over a list of vectors
}
