% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_number_of_clusters_stable_process.R
\name{asNumeric_no_warning}
\alias{asNumeric_no_warning}
\title{If the function Rmpfr::asNumeric returns a warning about inefficiency, silence it.}
\usage{
asNumeric_no_warning(x)
}
\arguments{
\item{x}{An object of class Rmpfr::mpfr1}
}
\value{
a "numeric" number
}
\description{
The function Rmpfr::asNumeric prints the following warning: In asMethod(object) : coercing "mpfr1" via "mpfr" (inefficient). It is not clear how to avoid it nor how to silence it, hence this function.
A cleaner solution may be available at: https://stackoverflow.com/questions/4948361/how-do-i-save-warnings-and-errors-as-output-from-a-function/4952908#4952908
}
