% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_code_rl.R
\name{censor_code_rl}
\alias{censor_code_rl}
\title{Censor code right-left}
\usage{
censor_code_rl(left, right)
}
\description{
Creates censoring code 0:3.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(left, right) {
  test_ <- function(k) {
    if (is.na(left[[k]]) & is.na(right[[k]])) {
      NA
    } else if (is.na(left[[k]])) {
      2
    } else if (is.na(right[[k]])) {
      0
    } else if (left[[k]] == right[[k]]) {
      1
    } else {
      3
    }
  }
  sapply(seq_along(left), FUN = test_)
}
}
\keyword{internal}
