% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_on_NRMI_type.R
\name{comment_on_NRMI_type}
\alias{comment_on_NRMI_type}
\title{Comment on the NRMI process depending on the value of the parameters}
\usage{
comment_on_NRMI_type(NRMI_param = list(Alpha = 1, Kappa = 0, Gamma = 0.4))
}
\arguments{
\item{NRMI_param}{A named list of the form list("Alpha" = 1, "Kappa" = 0, "Gamma" = 0.4)}
}
\value{
A string containing a comment on the NRMI process
}
\description{
Comment on the NRMI process depending on the value of the parameters
}
\examples{
BNPdensity:::comment_on_NRMI_type(list("Alpha" = 1, "Kappa" = 0, "Gamma" = 0.4))
BNPdensity:::comment_on_NRMI_type(list("Alpha" = 1, "Kappa" = 0.1, "Gamma" = 0.4))
BNPdensity:::comment_on_NRMI_type(list("Alpha" = 1, "Kappa" = 0.1, "Gamma" = 0.5))
}
