% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fit_to_MCMC_chain.R
\name{convert_to_mcmc}
\alias{convert_to_mcmc}
\title{Convert the output of multMixNRMI into a coda mcmc object}
\usage{
convert_to_mcmc(fitlist, thinning_to = 1000, ncores = parallel::detectCores())
}
\arguments{
\item{fitlist}{Output of multMixNRMI.}

\item{thinning_to}{Final length of the chain after thinning.}

\item{ncores}{Specify the number of cores to use in the conversion}
}
\value{
a coda::mcmc object
}
\description{
Convert the output of multMixNRMI into a coda mcmc object
}
