% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cpo.default}
\alias{cpo.default}
\title{Extract the Conditional Predictive Ordinates (CPOs) from a fitted object}
\usage{
\method{cpo}{default}(object, ...)
}
\arguments{
\item{object}{A fit obtained through one of the NRMI functions}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of Conditional Predictive Ordinates (CPOs)
}
\description{
Extract the Conditional Predictive Ordinates (CPOs) from a fitted object
}
