% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1.R
\name{cpo.NRMI1}
\alias{cpo.NRMI1}
\title{Extract the Conditional Predictive Ordinates (CPOs) from a fitted object}
\usage{
\method{cpo}{NRMI1}(object, ...)
}
\arguments{
\item{object}{A fit obtained through from the functions MixNRMI1/MixNRMI1cens}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A vector of Conditional Predictive Ordinates (CPOs)
}
\description{
Extract the Conditional Predictive Ordinates (CPOs) from a fitted object
}
\examples{
data(acidity)
out <- MixNRMI1(acidity, Nit = 50)
cpo(out)
}
