% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{cpo.NRMI2}
\alias{cpo.NRMI2}
\title{Extract the Conditional Predictive Ordinates (CPOs) from a fitted object}
\usage{
\method{cpo}{NRMI2}(object, ...)
}
\arguments{
\item{object}{A fit obtained through from the function MixNRMI2/MixNRMI2cens}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A vector of Conditional Predictive Ordinates (CPOs)
}
\description{
Extract the Conditional Predictive Ordinates (CPOs) from a fitted object
}
\examples{
data(acidity)
out <- MixNRMI2(acidity, Nit = 50)
cpo(out)
}
