% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpo.R, R/helper_functions.R
\name{cpo}
\alias{cpo}
\title{Conditional predictive ordinate function}
\usage{
cpo(object, ...)

cpo(object, ...)
}
\arguments{
\item{object}{A fit obtained through one of the NRMI functions}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of Conditional Predictive Ordinates (CPOs)
}
\description{
This function computes conditional predictive ordinates for each data point.
}
\details{
For internal use.
}
\examples{

## The function is currently defined as
function(obj) {
  fx <- obj$fx
  cpo <- 1 / apply(1 / fx, 1, mean)
  return(cpo)
}
}
\keyword{internal}
