% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_name_k_converter.R
\name{dist_name_k_index_converter}
\alias{dist_name_k_index_converter}
\title{Convert distribution names to indices}
\usage{
dist_name_k_index_converter(distname)
}
\arguments{
\item{distname}{a character representing the distribution name. Allowed names are "normal", "gamma", "beta", "exponential", "double exponential", "lognormal", "half-Cauchy", "half-normal", "half-student", "uniform" and "truncated normal", or their common abbreviations "norm", "exp", "lnorm", "halfcauchy", "halfnorm", "halft" and "unif".}
}
\value{
an index describing the distribution. 1 = Normal; 2
= Gamma; 3 = Beta; 4 = Double Exponential; 5 = Lognormal, 6 = Half-Cauchy, 7 = Half-normal, 8 = Half-Student, 9 = Uniform, 10 = Truncated normal
}
\description{
Convert distribution names to indices
}
