% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk.R
\name{dk}
\alias{dk}
\title{Kernel density function}
\usage{
dk(x, distr = NULL, mu = NULL, sigma = NULL)
}
\description{
This functions evaluates a density at a certain data point.  There are 4
density options (1 (normal), 2 (gamma), 3 (beta), 4 (exponential), 5 (lognormal), 6 (half-Cauchy), 7 (half-normal), 8 (half-student), 9 (uniform) and 10 (truncated normal)). All densities are parameterized in terms of mean and standard
deviation.
}
\details{
For internal use.
}
\keyword{internal}
