% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{enzyme}
\alias{enzyme}
\title{Enzyme Dataset}
\format{
A data frame with 244 observations on the following variable:
\describe{ \item{list("enzyme")}{A numeric vector.} }
}
\description{
Concerns the distribution of enzymatic activity in the blood, for an enzyme
involved in the metabolism of carcinogenetic substances, among a group of
245 unrelated individuals.
}
\examples{

data(enzyme)
hist(enzyme)
}
\references{
Bechtel, Y. C., Bonaiti-Pellie, C., Poisson, N., Magnette, J.
and Bechtel, P.R. (1993). A population and family study of
N-acetyltransferase using caffeine urinary metabolites. Clin. Pharm. Therp.,
54, 134-141.
}
\keyword{datasets}
