\name{Enzyme1.out}
\alias{Enzyme1.out}
\docType{data}
\title{
Fit of MixNRMI1 function to the enzyme dataset
}
\description{
This object contains the output when setting set.seed(123456) and running the function 
MixNRMI1(x, Alpha = 1, Beta = 0.007, Gama = 0.5, distr.k = 2, 
distr.p0 = 2, mu.p0 = 10, sigma.p0 = 10, asigma = 1, bsigma = 1,
Nit = 5000, Pbi = 0.2)
}
\usage{data(Enzyme1.out)}
\details{
See function MixNRMI1
}
\examples{
data(Enzyme1.out)
}
\keyword{datasets}
