\name{Enzyme2.out}
\alias{Enzyme2.out}
\docType{data}
\title{
Fit of MixNRMI2 function to the enzyme dataset
}
\description{
This object contains the output when setting set.seed(123456) and running the function
MixNRMI2(x, Alpha = 1, Beta = 0.007, Gama = 0.5, distr.k = 2, 
distr.py0 = 2, mu.py0 = 10, sigma.py0 = 10,
distr.pz0 = 2, mu.pz0 = 1, sigma.pz0 = 1,
Nit = 5000, Pbi = 0.2)
}
\usage{data(Enzyme2.out)}
\details{
See function MixNRMI2
}
\examples{
data(Enzyme2.out)
}
\keyword{datasets}
