% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcondYXA.R
\name{fcondYXA}
\alias{fcondYXA}
\title{Conditional posterior distribution of the latents Y}
\usage{
fcondYXA(x, distr, Tau, J, sigma)
}
\description{
This function simulates form the conditional posterior distribution of the
latents Y.
}
\details{
For internal use.
}
\examples{

## The function is currently defined as
function(x, distr = 1, Tau, J, sigma) {
  K <- matrix(NA, nrow = length(Tau), ncol = length(x))
  for (i in seq(Tau)) {
    K[i, ] <- dk(x, distr = distr, mu = Tau[i], sigma = sigma) *
      J[i]
  }
  pK <- prop.table(K, margin = 2)
  y <- apply(pK, 2, function(x) sample(Tau, size = 1, prob = x))
  return(y)
}
}
\keyword{internal}
