% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy Data Set}
\format{
A data frame with 82 observations on the following variable:
\describe{ \item{list("velocity")}{A numeric vector.} }
}
\description{
Velocities of 82 galaxies diverging from our own galaxy.
}
\examples{

data(galaxy)
hist(galaxy)
}
\references{
Roeder, K. (1990) "Density estimation with confidence sets
exemplified by superclusters and voids in the galaxies". Journal of the
American Statistical Association. 85, 617-624.
}
\keyword{datasets}
