\name{Galaxy2.out}
\alias{Galaxy2.out}
\docType{data}
\title{
Fit of MixNRMI2 function to the galaxy dataset
}
\description{
This object contains the output when setting set.seed(123456) and running the function 
MixNRMI2(x, Alpha = 1, Beta = 0.015, Gama = 0.5, distr.k = 1, 
distr.py0 = 2, mu.py0 = 20, sigma.py0 = 20,
distr.pz0 = 2, mu.pz0 = 1, sigma.pz0 = 1,
Nit = 5000, Pbi = 0.2)
}
\usage{data(Galaxy2.out)}
\details{
See function MixNRMI2.
}
\examples{
data(Galaxy2.out)
}
\keyword{datasets}
