% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_kernel_name.R
\name{give_kernel_name}
\alias{give_kernel_name}
\title{Gives the kernel name from the integer code}
\usage{
give_kernel_name(distr.k)
}
\arguments{
\item{distr.k}{The distribution name for the kernel. Allowed names are "normal", "gamma", "beta", "double exponential", "lognormal" or their common abbreviations "norm", "exp", or an integer number identifying the mixture kernel: 1 = Normal; 2 = Gamma; 3 = Beta; 4 = Double Exponential; 5 = Lognormal.}
}
\value{
A character with the name of the distribution used as the kernel
}
\description{
This function is used in the print methods for MixNRMI1, MixNRMI2, MixNRMI1cens, MixNRMI2cens, and all the multMixNRMIx versions
}
\examples{
BNPdensity:::give_kernel_name(4)
}
