% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFplots.R
\name{GOFplots}
\alias{GOFplots}
\title{Plot Goodness of fits graphical checks for censored data}
\usage{
GOFplots(fit, qq_plot = FALSE, thinning_to = 500)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or
MixNRMI2, MixMRMI1cens or MixMRMI2cens}

\item{qq_plot}{Whether to compute the QQ-plot}

\item{thinning_to}{How many iterations to compute the mean posterior
quantiles}
}
\value{
A density plot, a cumulative density plot with the Turnbull
cumulative distribution, a percentile-percentile plot, and potentially a
quantile-quantile plot.
}
\description{
Plot Goodness of fits graphical checks for censored data
}
\examples{

set.seed(150520)
data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, extras = TRUE, Nit = 100)
GOFplots(out)
}
