% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{grid_from_data}
\alias{grid_from_data}
\title{Create a plotting grid from censored or non-censored data.}
\usage{
grid_from_data(data, npoints = 100)
}
\arguments{
\item{data}{Input data from which to compute the grid.}

\item{npoints}{Number of points on the grid.}
}
\value{
a vector containing the plotting grid
}
\description{
Create a plotting grid from censored or non-censored data.
}
