% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{grid_from_data_censored}
\alias{grid_from_data_censored}
\title{Create a plotting grid from censored data.}
\usage{
grid_from_data_censored(data, npoints = 100)
}
\arguments{
\item{data}{Censored input data from which to compute the grid.}

\item{npoints}{Number of points on the grid.}
}
\value{
a vector containing the plotting grid
}
\description{
Create a plotting grid from censored data.
}
