% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{gs3_adaptive3}
\alias{gs3_adaptive3}
\title{Conditional posterior distribution of latent U}
\usage{
gs3_adaptive3(ut, n, r, alpha, kappa, gama, delta, U, iter, adapt = FALSE)
}
\description{
This function simulates from the conditional posterior distribution of the
latent U, with an adaptive proposal
}
\keyword{internal}
