% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{gs3_log}
\alias{gs3_log}
\title{Conditional posterior distribution of latent logU}
\usage{
gs3_log(logut, n, r, alpha, beta, gama, delta)
}
\description{
This function simulates from the conditional posterior distribution of a log transformation of the
latent U.
}
