% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_censored}
\alias{is_censored}
\title{Test if the data is censored}
\usage{
is_censored(dat)
}
\arguments{
\item{dat}{The dataset to be tested}
}
\value{
TRUE if the data is censored
}
\description{
Test if the data is censored
}
\examples{

data(salinity)
BNPdensity:::is_censored(salinity)
}
