% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_semiparametric}
\alias{is_semiparametric}
\title{Tests if a fit is a semi parametric or nonparametric model.}
\usage{
is_semiparametric(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or
MixNRMI2.}
}
\value{
TRUE if the fit is a semiparametric model
}
\description{
Tests if a fit is a semi parametric or nonparametric model.
}
\examples{

set.seed(150520)
data(acidity)
x <- enzyme
out <- MixNRMI1(enzyme, extras = TRUE, Nit = 10)
BNPdensity:::is_semiparametric(out)
}
