% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{logdprop_logu}
\alias{logdprop_logu}
\title{Contribution of the proposal kernel logdensity to the Metropolis-Hastings ratio}
\usage{
logdprop_logu(logu_prime, logu, delta)
}
\description{
Contribution of the proposal kernel logdensity to the Metropolis-Hastings ratio
}
\keyword{internal}
