% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{logf_logu_cond_y}
\alias{logf_logu_cond_y}
\title{Contribution of the target logdensity of logU to the Metropolis-Hastings ratio}
\usage{
logf_logu_cond_y(logu, n, r, gamma, kappa, a)
}
\description{
Contribution of the target logdensity of logU to the Metropolis-Hastings ratio
}
\keyword{internal}
