% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixPY1.R
\name{MixPY1}
\alias{MixPY1}
\title{Title}
\usage{
MixPY1(
  x,
  Alpha = 1,
  Gama = 0.4,
  Nx = 150,
  Nit = 1500,
  Pbi = 0.5,
  epsilon = NULL,
  printtime = TRUE
)
}
\arguments{
\item{printtime}{}
}
\value{

}
\description{
Title
}
