% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvInv.R
\name{MvInv}
\alias{MvInv}
\title{Invert jump heights function}
\usage{
MvInv(eps, u = 0.5, alpha = 1, beta = 1, gama = 1/2, N = 3001)
}
\arguments{
\item{eps}{Dummy argument kept for consistency with past versions of the functions}

\item{u}{Real number. The value of the latent variable at the current step.}

\item{alpha}{Numeric constant. Total mass of the centering measure.}

\item{beta}{Numeric positive constant.}

\item{gama}{Numeric constant. Discount parameter of the NRMI process.}
}
\description{
Determines the jump heights of an increasing additive process by inverting
the M(v) function. Use a truncation level based on expected moments of the NGG process (\code{\link{thresholdGG}}).
For internal use.
}
