% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phalfcauchy.R
\name{phalfcauchy}
\alias{phalfcauchy}
\title{Distribution function half Cauchy}
\usage{
phalfcauchy(q, location = 0, scale = 1)
}
\description{
Computes the cdf.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(q, location = 0, scale = 1) {
  ifelse(x < 0, 0, 1) * (pcauchy(q, location, scale) - pcauchy(
    0,
    location, scale
  )) / (1 - pcauchy(0, location, scale))
}
}
\keyword{internal}
