% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phalft.R
\name{phalft}
\alias{phalft}
\title{Distribution function half Student-t}
\usage{
phalft(q, df = 1, mean = 0, sd = 1)
}
\description{
Computes the cumulative distribution function.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(q, df = 1, mean = 0, sd = 1) {
  ifelse(x < 0, 0, 1) * (pt_(q, df, mean, sd) - pt_(
    0, df,
    mean, sd
  )) / (1 - pt_(0, df, mean, sd))
}
}
\keyword{internal}
