% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1cens.R
\name{plot.NRMI1cens}
\alias{plot.NRMI1cens}
\title{Plot the density estimate and the 95\% credible interval}
\usage{
\method{plot}{NRMI1cens}(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI1cens}
}
\value{
A graph with the density estimate, the 95\% credible interval
}
\description{
The density estimate is the mean posterior density computed on the data
points. It is not possible to display a histogram for censored data.
}
\examples{

data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, Nit = 50)
plot(out)
}
