% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{plot.NRMI2}
\alias{plot.NRMI2}
\title{Plot the density estimate and the 95\% credible interval}
\usage{
\method{plot}{NRMI2}(x, ...)
}
\arguments{
\item{x}{A fitted object of class NRMI2}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A graph with the density estimate, the 95\% credible interval and a
histogram of the data
}
\description{
The density estimate is the mean posterior density computed on the data
points.
}
\examples{
## Example for non censored data

data(acidity)
out <- MixNRMI2(acidity, Nit = 20)
plot(out)

## Example for censored data

data(salinity)
out <- MixNRMI2cens(salinity$left, salinity$right, Nit = 20)
plot(out)
}
