% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixPY1.R
\name{plot.PY1}
\alias{plot.PY1}
\title{Plot the density estimate and the 95\% credible interval}
\usage{
\method{plot}{PY1}(x, ...)
}
\arguments{
\item{x}{A fitted object of class PY1}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A graph with the density estimate, the 95\% credible interval and a
histogram of the data
}
\description{
Plot the density estimate and the 95\% credible interval
}
\examples{
data(acidity)
out <- MixPY1(acidity, Nit = 50)
plot(out)
}
