% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFplots.R
\name{plotCDF_noncensored}
\alias{plotCDF_noncensored}
\title{Plot the empirical and fitted CDF for non censored data.}
\usage{
plotCDF_noncensored(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or
MixNRMI2.}
}
\value{
Plot of the empirical and fitted CDF for non censored data.
}
\description{
Plot the empirical and fitted CDF for non censored data.
}
\examples{

set.seed(150520)
data(acidity)
out <- MixNRMI1(acidity, extras = TRUE, Nit = 10)
BNPdensity:::plotCDF_noncensored(out)
}
