% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R
\name{plotfit_censored}
\alias{plotfit_censored}
\title{Plot the density estimate and the 95\% credible interval for censored data}
\usage{
plotfit_censored(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI1cens or NRMI2cens}
}
\value{
A graph with the density estimate and the 95\% credible interval
}
\description{
The density estimate is the mean posterior density computed on the data
points. It is not possible to display a histogram for censored data.
}
\examples{

data(acidity)
out <- MixNRMI1(acidity, Nit = 50)
plot(out)
}
