% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R
\name{plotfit_manylines}
\alias{plotfit_manylines}
\title{Plot the density estimate and grey lines to represent the uncertainty on the density estimate}
\usage{
plotfit_manylines(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI1 or NRMI2}
}
\value{
A graph with the density estimate, the 95\% credible interval and a histogram of the data.
}
\description{
Plot the density estimate and grey lines to represent the uncertainty on the density estimate
}
