% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_plots.R
\name{plotGOF}
\alias{plotGOF}
\title{Plot Goodness of fits graphical checks for censored data}
\usage{
plotGOF(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or MixNRMI2, MixMRMI1cens or MixMRMI2cens}
}
\value{
A density plot, a cumulative density plot with the Turnbull cumulative distribution, and a percentile-percentile plot.
}
\description{
Plot Goodness of fits graphical checks for censored data
}
\examples{
set.seed(150520)
data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, extras = TRUE, Nit = 100)
plotGOF(out)
}
