% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFplots.R
\name{plotPDF_censored}
\alias{plotPDF_censored}
\title{Plot the density for censored data.}
\usage{
plotPDF_censored(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1cens
or MixNRMI2cens.}
}
\value{
Plot of the density and a histogram for non censored data.
}
\description{
Plot the density for censored data.
}
\examples{

set.seed(150520)
data(salinity)
out <- MixNRMI1cens(xleft = salinity$left, xright = salinity$right, extras = TRUE, Nit = 100)
BNPdensity:::plotPDF_censored(out)
}
