% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFplots.R
\name{plotPDF_noncensored}
\alias{plotPDF_noncensored}
\title{Plot the density and a histogram for non censored data.}
\usage{
plotPDF_noncensored(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or
MixNRMI2.}
}
\value{
Plot of the density and a histogram for non censored data.
}
\description{
Plot the density and a histogram for non censored data.
}
\examples{

set.seed(150520)
data(acidity)
out <- MixNRMI1(acidity, extras = TRUE, Nit = 100)
BNPdensity:::plotPDF_noncensored(out)
}
