% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFplots.R
\name{pp_plot_noncensored}
\alias{pp_plot_noncensored}
\title{Plot the percentile-percentile graph for non censored data.}
\usage{
pp_plot_noncensored(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1 or
MixNRMI2.}
}
\value{
Percentile-percentile plot for non censored data.
}
\description{
Plot the percentile-percentile graph for non censored data.
}
\examples{

set.seed(150520)
data(acidity)
out <- MixNRMI1(acidity, extras = TRUE, Nit = 100)
BNPdensity:::pp_plot_noncensored(out)
}
