% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multMixNRMI.R
\name{print.multNRMI}
\alias{print.multNRMI}
\title{S3 method for class 'multNRMI'}
\usage{
\method{print}{multNRMI}(fitlist)
}
\arguments{
\item{fitlist}{An object of class multNRMI}
}
\value{
A visualisation of the important information about the object
}
\description{
S3 method for class 'multNRMI'
}
\examples{
data(salinity)
out <- multMixNRMI2cens(salinity$left, salinity$right, parallel = TRUE, Nit = 20, ncores = 2)
print(out)
}
