% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1cens.R
\name{print.NRMI1cens}
\alias{print.NRMI1cens}
\title{S3 method for class 'MixNRMI1cens'}
\usage{
\method{print}{NRMI1cens}(x, ...)
}
\arguments{
\item{x}{A fitted object of class MixNRMI1cens}

\item{...}{Further arguments to be passed to generic functions, ignored at the moment}
}
\value{
A visualization of the important information about the object
}
\description{
S3 method for class 'MixNRMI1cens'
}
\examples{
data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, Nit = 50)
print(out)
}
