% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{print.NRMI2}
\alias{print.NRMI2}
\title{S3 method for class 'MixNRMI2'}
\usage{
\method{print}{NRMI2}(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI2}
}
\value{
A visualisation of the important information about the object
}
\description{
S3 method for class 'MixNRMI2'
}
\examples{
data(acidity)
out <- MixNRMI2(acidity, Nit = 50)
print(out)
}
