% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixPY1.R
\name{print.PY1}
\alias{print.PY1}
\title{S3 method for class 'PY1'}
\usage{
\method{print}{PY1}(x, ...)
}
\arguments{
\item{x}{A fitted object of class PY1}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A visualization of the important information about the object
}
\description{
S3 method for class 'PY1'
}
\examples{

## Example for non censored data

data(acidity)
out <- MixPY1(acidity, Nit = 50)
print(out)
}
