% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixPY2.R
\name{print.PY2}
\alias{print.PY2}
\title{S3 method for class 'PY2'}
\usage{
\method{print}{PY2}(x, ...)
}
\arguments{
\item{x}{A fitted object of class PY2}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
A visualization of the important information about the object
}
\description{
S3 method for class 'PY2'
}
\examples{

## Example for non censored data

data(acidity)
out <- MixPY2(acidity, Nit = 50)
print(out)
}
