% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_name_k_converter.R
\name{process_dist_name}
\alias{process_dist_name}
\title{Process the distribution name argument into a distribution index}
\usage{
process_dist_name(distname)
}
\arguments{
\item{distname}{Can be an integer or a distribution name. Allowed names are "normal", "gamma", "beta", "exponential", "lognormal", "half-Cauchy", "half-normal", "half-student", "uniform" and "truncated normal", or their common abbreviations "norm", "exp", "halfcauchy", "halfnorm", "halft" and "unif".}
}
\value{
an integer both if distname is an integer or a character
}
\description{
This function is intended to help with compatibility with the previous versions of the package.
}
