% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_.R
\name{qt_}
\alias{qt_}
\title{Quantile function non-standard Student-t}
\usage{
qt_(p, df, mean, sd)
}
\description{
Computes the quantiles.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(p, df, mean, sd) {
  sd * qt(p, df, ncp = 0) + mean
}
}
\keyword{internal}
