% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtnorm.R
\name{qtnorm}
\alias{qtnorm}
\title{Quantile function truncated normal}
\usage{
qtnorm(
  p,
  mean = 0,
  sd = 1,
  lower = -Inf,
  upper = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\description{
Computes the quantiles.
}
\details{
For internal use
}
\note{
Taken from \code{msm} R-package.
}
\examples{

## The function is currently defined as
function(p, mean = 0, sd = 1, lower = -Inf, upper = Inf, lower.tail = TRUE,
         log.p = FALSE) {
  qgeneric(ptnorm,
    p = p, mean = mean, sd = sd, lower = lower,
    upper = upper, lbound = lower, ubound = upper, lower.tail = lower.tail,
    log.p = log.p
  )
}
}
\references{
Taken from
}
\author{
C. H. Jackson
}
\keyword{internal}
