% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhalfcauchy.R
\name{rhalfcauchy}
\alias{rhalfcauchy}
\title{Random number generator half Cauchy}
\usage{
rhalfcauchy(n, location = 0, scale = 1)
}
\description{
Computes a random number.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(n, location = 0, scale = 1) {
  abs(rcauchy(n, location, scale))
}
}
\keyword{internal}
