% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhalfnorm.R
\name{rhalfnorm}
\alias{rhalfnorm}
\title{Random number generator half Normal}
\usage{
rhalfnorm(n, mean = 0, sd = 1)
}
\description{
Computes a random number.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(n, mean = 0, sd = 1) {
  abs(rnorm(n, mean, sd))
}
}
\keyword{internal}
