% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{rprop_logu}
\alias{rprop_logu}
\title{Proposal distribution for logU}
\usage{
rprop_logu(logu, delta)
}
\arguments{
\item{logu}{Real, log of the latent variable U at the current iteration.}

\item{delta}{Scale of the Metropolis-Hastings proposal distribution}
}
\description{
This function makes a proposal for a new value of logU
}
\keyword{internal}
