% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1.R
\name{summary.NRMI1}
\alias{summary.NRMI1}
\title{S3 method for class 'MixNRMI1'}
\usage{
\method{summary}{NRMI1}(object, number_of_clusters = FALSE, ...)
}
\arguments{
\item{object}{A fitted object of class NRMI1}

\item{number_of_clusters}{Whether to compute the optimal number of clusters, which can be a time-consuming operation (see \code{\link{compute_optimal_clustering}})}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'MixNRMI1'
}
\examples{

## Example for non censored data

data(acidity)
out <- MixNRMI1(acidity, Nit = 50)
summary(out)
}
