% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1cens.R
\name{summary.NRMI1cens}
\alias{summary.NRMI1cens}
\title{S3 method for class 'MixNRMI1cens'}
\usage{
\method{summary}{NRMI1cens}(object, number_of_clusters = FALSE, ...)
}
\arguments{
\item{object}{A fitted object of class NRMI1cens}

\item{number_of_clusters}{Whether to compute the optimal number of clusters, which can be a time-consuming operation (see \code{\link{compute_optimal_clustering}})}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
Summary of some relevant information on the fit
}
\description{
S3 method for class 'MixNRMI1cens'
}
\examples{
data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, Nit = 50)
summary(out)
}
