% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{summary.NRMI2}
\alias{summary.NRMI2}
\title{S3 method for class 'MixNRMI2'}
\usage{
\method{summary}{NRMI2}(fit, number_of_clusters = FALSE)
}
\arguments{
\item{fit}{A fitted object of class NRMI2}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'MixNRMI2'
}
\examples{
data(acidity)
out <- MixNRMI2(acidity, Nit = 50)
summary(out)
}
