% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixPY1.R
\name{summary.PY1}
\alias{summary.PY1}
\title{S3 method for class 'PY1'}
\usage{
\method{summary}{PY1}(object, number_of_clusters = FALSE, ...)
}
\arguments{
\item{object}{A fitted object of class PY1}

\item{number_of_clusters}{Whether to compute the optimal number of clusters, which can be a time-consuming operation (see \code{\link{compute_optimal_clustering}})}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'PY1'
}
\examples{

## Example for non censored data

data(acidity)
out <- MixPY1(acidity, Nit = 50)
summary(out)
}
