% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytext.R
\name{summarytext}
\alias{summarytext}
\title{Common text for the summary S3 methods}
\usage{
summarytext(fit, kernel_comment, number_of_clusters = FALSE)
}
\arguments{
\item{fit}{NRMIx object}

\item{kernel_comment}{Text specific to the parametric and nonparametric nature of the model}

\item{number_of_clusters}{Flag to decide whether to compute the optimal clustering}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
Common text for the summary S3 methods
}
