% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdGG.R
\name{thresholdGG}
\alias{thresholdGG}
\title{Choosing the truncation level for the NGG process}
\usage{
thresholdGG(alpha = 1, kappa = 1, gama = 1/2, max_threshold = 200)
}
\arguments{
\item{alpha}{Numeric constant. Total mass of the centering measure}

\item{kappa}{Numeric positive constant.}

\item{gama}{Numeric constant. \eqn{0 \leq Gama \leq 1}{0 <= Gama <=1}.}

\item{max_threshold}{Numeric positive integer. Maximum allowed value for the threshold}
}
\value{
Numeric positive integer, the truncation level of the NGG process
}
\description{
This function uses the M_array which provides the threshold which ensures
a moment match of 5% for NGG parameters alpha, kappa, gama
}
\details{
For internal use
}
\keyword{internal}
