% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{acidity}
\alias{acidity}
\title{Acidity Index Dataset}
\format{A data frame with 155 observations on the following variable:
\describe{ \item{list("acidity")}{A numeric vector.} }}
\description{
Concerns an acidity index measured in a sample of 155 lakes in north-central
Wisconsin.
}
\examples{

data(acidity)
hist(acidity)
}
\references{
Crawford, S. L., DeGroot, M. H., Kadane, J. B. and Small, M. J.
(1992). Modeling lake chemistry distributions: approximate Bayesian methods
for estimating a finite mixture model. Technometrics, 34, 441-453.
}
\keyword{datasets}
