% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cens_data_check.R
\name{cens_data_check}
\alias{cens_data_check}
\title{Censoring data check}
\usage{
cens_data_check(xleft, xright)
}
\description{
Performs censoring data check.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function (xleft, xright) 
{
    if (any(xright < xleft, na.rm = T)) 
        stop("in censored data, left bound not always smaller than right bound")
    if (any(mapply(FUN = function(xileft, xiright) {
        is.na(xileft) & is.na(xiright)
    }, xleft, xright))) 
        stop("in censored data, there is an NA NA")
  }

}
\keyword{internal}
