% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_clustering_analysis.R
\name{compute_optimal_clustering}
\alias{compute_optimal_clustering}
\title{Compute the optimal clustering from an MCMC sample}
\usage{
compute_optimal_clustering(fit, loss_function = "VI", silent = FALSE)
}
\arguments{
\item{fit}{The fitted object, obtained from one of the MixNRMIx functions}

\item{loss_function}{Choice of the loss function. Can be "VI" or "Binder"}

\item{silent}{Whether to silence the warning message about the length of the calculation.}
}
\value{
A vector of integers with the same size as the data, indicating the allocation of each data point.
}
\description{
Summarises the posterior on all possible clusterings by an optimal clustering where optimality is defined as minimising the posterior expectation of a specific loss function, the Variation of Information or Binder's loss function.
Computation can be lengthy for large datasets, because of the large size of the space of all clusterings.
}
