% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpo.R
\name{cpo}
\alias{cpo}
\title{Conditional predictive ordinate function}
\usage{
cpo(obj)
}
\description{
This function computes conditional predictive ordinates for each datapoint.
}
\details{
For internal use.
}
\examples{

## The function is currently defined as
function (obj) 
{
    fx <- obj$fx
    cpo <- 1/apply(1/fx, 1, mean)
    return(cpo)
  }

}
\keyword{internal}
