% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhalft.R
\name{dhalft}
\alias{dhalft}
\title{Density half Student-t}
\usage{
dhalft(x, df = 1, mean = 0, sd = 1)
}
\description{
Computes the density.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(x, df = 1, mean = 0, sd = 1) {
  ifelse(x < 0, 0, 1) * dt_(x, df, mean, sd) / (1 - pt_(
    0, df,
    mean, sd
  ))
}
}
\keyword{internal}
