% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_.R
\name{dt_}
\alias{dt_}
\title{Non-standard student-t density}
\usage{
dt_(x, df, mean, sd)
}
\arguments{
\item{x}{Numeric vector. Data set to which the density is evaluated.}

\item{df}{Numeric constant. Degrees of freedom (> 0, maybe non-integer)}

\item{mean}{Numeric constant. Location parameter.}

\item{sd}{Positive numeric constant. Scale parameter.}
}
\description{
Computes the density.
}
\details{
For internal use
}
