% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_.R
\name{dt_}
\alias{dt_}
\title{Non-standard student-t densit}
\usage{
dt_(x, df, mean, sd)
}
\description{
Computes the density.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function (x, df, mean, sd) 
{
    dt((x - mean)/sd, df, ncp = 0)/sd
  }

}
\keyword{internal}
