% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_.R
\name{dt_}
\alias{dt_}
\title{Non-standard student-t density}
\usage{
dt_(x, df, mean, sd)
}
\description{
Computes the density.
}
\details{
For internal use
}
\keyword{##}
\keyword{-}
\keyword{/}
\keyword{0)}
\keyword{=}
\keyword{The}
\keyword{as}
\keyword{currently}
\keyword{defined}
\keyword{df,}
\keyword{dt((x}
\keyword{function}
\keyword{function(x,}
\keyword{internal}
\keyword{is}
\keyword{mean)}
\keyword{mean,}
\keyword{ncp}
\keyword{sd}
\keyword{sd)}
\keyword{sd,}
\keyword{{}
\keyword{}}
