% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{Enzyme2.out}
\alias{Enzyme2.out}
\title{Fit of MixNRMI2 function to the enzyme dataset}
\description{
This object contains the output when setting set.seed(150520) and running
the function Enzyme2.out <- MixNRMI2(x, Alpha = 1, Kappa = 0.007, Gama =
0.5, distr.k = 2, distr.py0 = 2, distr.pz0 = 2, mu.pz0 = 1, sigma.pz0 = 1,
Meps=0.001, Nit = 5000, Pbi = 0.2)
}
\details{
See function MixNRMI2
}
\examples{

data(Enzyme2.out)

}
\keyword{datasets}
