% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{Galaxy1.out}
\alias{Galaxy1.out}
\title{Fit of MixNRMI1 function to the galaxy dataset}
\description{
This object contains the output when setting set.seed(150520) and running
the function MixNRMI1(galaxy, Alpha = 1, Kappa = 0.015, Gama = 0.5, distr.k = "normal", distr.p0 = "gamma", asigma = 1, bsigma = 1, delta = 7, Meps = 0.005, Nit = 5000, Pbi = 0.2)
}
\details{
See function MixNRMI1.
}
\examples{

data(Galaxy1.out)
}
\keyword{datasets}
