\name{Galaxy1.out}
\alias{Galaxy1.out}
\docType{data}
\title{
Fit of MixNRMI1 function to the galaxy dataset
}
\description{
This object contains the output when setting set.seed(150520) and running the function 
MixNRMI1(x, Alpha = 1, Beta = 0.015, Gama = 0.5, distr.k = 1, distr.p0 = 2, 
  asigma = 1, bsigma = 1,  Meps=0.001, Nit = 5000, Pbi = 0.2)
}
\usage{data(Galaxy1.out)}
\details{
See function MixNRMI1.
}
\examples{
data(Galaxy1.out)
}
\keyword{datasets}
