% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{Galaxy2.out}
\alias{Galaxy2.out}
\title{Fit of MixNRMI2 function to the galaxy dataset}
\description{
This object contains the output when setting set.seed(150520) and running
the function Enzyme2.out <- MixNRMI2(x, Alpha = 1, Kappa = 0.007, Gama = 0.5, distr.k = "gamma", distr.py0 = "gamma", distr.pz0 = "gamma", mu.pz0 = 1, sigma.pz0 = 1, Meps = 0.005, Nit = 5000, Pbi = 0.2)
}
\details{
See function MixNRMI2.
}
\examples{

data(Galaxy2.out)
}
\keyword{datasets}
