% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs3.R
\name{logacceptance_ratio_logu}
\alias{logacceptance_ratio_logu}
\title{Metropolis-Hastings ratio for the conditional of logU}
\usage{
logacceptance_ratio_logu(logu, logu_prime, n, r, gamma, kappa, a, delta)
}
\description{
This function computes the Metropolis-Hastings ratio to decide whether to accept or reject a new value for logU.
}
\details{
@keywords internal
}
