% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multMixNRMI.R
\name{multMixNRMI1}
\alias{multMixNRMI1}
\title{Multiple chains of MixNRMI1}
\usage{
multMixNRMI1(x, probs = c(0.025, 0.5, 0.975), Alpha = 1, Beta = 0,
  Gama = 0.4, distr.k = 1, distr.p0 = 1, asigma = 0.5,
  bsigma = 0.5, delta = 3, Delta = 2, Meps = 0.01, Nx = 150,
  Nit = 1500, Pbi = 0.1, epsilon = NULL, printtime = TRUE,
  extras = FALSE, nchains = 4, parallel = TRUE, seed = 1)
}
\arguments{
\item{x}{Numeric vector. Data set to which the density is fitted.}

\item{probs}{Numeric vector. Desired quantiles of the density estimates.}

\item{Alpha}{Numeric constant. Total mass of the centering measure. See details.}

\item{Beta}{Numeric positive constant. See details.}

\item{Gama}{Numeric constant. 0 <= Gama <= 1. See details.}

\item{distr.k}{Integer number identifying the mixture kernel: 1 = Normal; 2 = Gamma; 3 = Beta; 4 = Double Exponential; 5 = Lognormal.}

\item{distr.p0}{Integer number identifying the centering measure: 1 = Normal; 2 = Gamma; 3 = Beta.}

\item{asigma}{Numeric positive constant. Shape parameter of the gamma prior on the standard deviation of the mixture kernel distr.k.}

\item{bsigma}{Numeric positive constant. Rate parameter of the gamma prior on the standard deviation of the mixture kernel distr.k.}

\item{delta}{Numeric positive constant. Metropolis-Hastings proposal variation coefficient for sampling sigma.}

\item{Delta}{Numeric positive constant. Metropolis-Hastings proposal variation coefficient for sampling the latent U.}

\item{Meps}{Numeric constant. Relative error of the jump sizes in the continuous component of the process. Smaller values imply larger number of jumps.}

\item{Nx}{Integer constant. Number of grid points for the evaluation of the density estimate.}

\item{Nit}{Integer constant. Number of MCMC iterations.}

\item{Pbi}{Numeric constant. Burn-in period proportion of Nit.}

\item{epsilon}{Numeric constant. Extension to the evaluation grid range. See details.}

\item{printtime}{Logical. If TRUE, prints out the execution time.}

\item{extras}{Logical. If TRUE, gives additional objects: means, weights and Js (the jump sizes).}

\item{nchains}{The number of chains to run.}

\item{parallel}{Whether to run the chains in parallel. Only works on UNIX-like systems as it rests on Fork parallelism}

\item{seed}{Random seed for the run. Each chain has an independent seed based on the value of `seed`}
}
\value{
a list containing the multiple fits.
}
\description{
Multiple chains of MixNRMI1
}
\examples{
data(acidity)
MultMixNRMI1(acidity, parallel = TRUE, Nit = 100)
}
