% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvInv.R
\name{MvInv}
\alias{MvInv}
\title{Invert jump heights function}
\usage{
MvInv(eps, u = 0.5, alpha = 1, beta = 1, gama = 1/2, N = 3001)
}
\description{
Determines the jump heights of an increasing additive process by inverting
the M(v) function.
}
\details{
For internal use.
}
\keyword{##}
\keyword{(f[-1]}
\keyword{(i}
\keyword{(j}
\keyword{*}
\keyword{+}
\keyword{-}
\keyword{-log(seq(from}
\keyword{/}
\keyword{0.5,}
\keyword{1}
\keyword{1))}
\keyword{1,}
\keyword{1]}
\keyword{2}
\keyword{2,}
\keyword{3001)}
\keyword{<-}
\keyword{=}
\keyword{>}
\keyword{Mv}
\keyword{Mv[i}
\keyword{Mv[i]}
\keyword{N}
\keyword{N)}
\keyword{N))}
\keyword{The}
\keyword{alpha}
\keyword{as}
\keyword{been}
\keyword{beta}
\keyword{beta)}
\keyword{but}
\keyword{defined}
\keyword{diff(x)}
\keyword{dx}
\keyword{dx[i]}
\keyword{exp(-(u}
\keyword{exp(-10),}
\keyword{exp(-1e-05),}
\keyword{f}
\keyword{f[-N])}
\keyword{for}
\keyword{function}
\keyword{function(eps,}
\keyword{gama}
\keyword{gama)}
\keyword{gama))}
\keyword{gamma(1}
\keyword{h}
\keyword{h[i]}
\keyword{has}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{length}
\keyword{length(w)}
\keyword{morally}
\keyword{n}
\keyword{n)}
\keyword{optimised,}
\keyword{rep(0,}
\keyword{rep(NA,}
\keyword{return(v)}
\keyword{seq(N}
\keyword{seq(n))}
\keyword{to}
\keyword{u}
\keyword{v}
\keyword{v[j]}
\keyword{w[j])]}
\keyword{x}
\keyword{x)}
\keyword{x[which.min(Mv}
\keyword{x^(-(1}
\keyword{{}
\keyword{}}
