% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multMixNRMI.R
\name{plot.multNRMI}
\alias{plot.multNRMI}
\title{Plot the density estimate and the 95\% credible interval}
\usage{
\method{plot}{multNRMI}(fitlist)
}
\arguments{
\item{fitlist}{A fitted object of class multNRMI}
}
\value{
A graph with the density estimate, the 95\% credible interval.
Includes a histogram if the data is non censored.
}
\description{
The density estimate is the mean posterior density computed on the data
points.
}
\examples{

fit <- multMixNRMI2cens(salinity$left, salinity$right, parallel = TRUE, Nit = 20, ncores = 2)
plot(fit)
}
