% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{plot.NRMI2}
\alias{plot.NRMI2}
\title{Plot the density estimate and the 95\% credible interval}
\usage{
\method{plot}{NRMI2}(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI2}
}
\value{
A graph with the density estimate, the 95\% credible interval and a
histogram of the data
}
\description{
The density estimate is the mean posterior density computed on the data
points.
}
\examples{

data(acidity)
out <- MixNRMI2(acidity, Nit = 50)
plot(out)
}
