% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multMixNRMI.R
\name{print.multNRMI}
\alias{print.multNRMI}
\title{S3 method for class 'multNRMI'}
\usage{
\method{print}{multNRMI}(x, ...)
}
\arguments{
\item{x}{An object of class multNRMI}

\item{...}{Further arguments to be passed to generic functions, ignored at the moment}
}
\value{
A visualization of the important information about the object
}
\description{
S3 method for class 'multNRMI'
}
\examples{
\donttest{
data(salinity)
out <- multMixNRMI2cens(salinity$left, salinity$right, parallel = TRUE, Nit = 10, ncores = 2)
print(out)
}
}
