% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1.R
\name{print.NRMI1}
\alias{print.NRMI1}
\title{S3 method for class 'MixNRMI1'}
\usage{
\method{print}{NRMI1}(fit)
}
\arguments{
\item{fit}{A fitted object of class NRMI1}
}
\value{
A visualisation of the important information about the object
}
\description{
S3 method for class 'MixNRMI1'
}
\examples{
data(acidity)
out <- MixNRMI1(acidity, Nit = 50)
print(out)
}
