% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2cens.R
\name{print.NRMI2cens}
\alias{print.NRMI2cens}
\title{S3 method for class 'MixNRMI2cens'}
\usage{
\method{print}{NRMI2cens}(x, ...)
}
\arguments{
\item{x}{A fitted object of class MixNRMI2cens}

\item{...}{Further arguments to be passed to generic functions, ignored at the moment}
}
\value{
A visualisation of the important information about the object
}
\description{
S3 method for class 'MixNRMI2cens'
}
\examples{
data(salinity)
out <- MixNRMI2cens(salinity$left, salinity$right, Nit = 50)
print(out)
}
