% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhalfnorm.R
\name{qhalfnorm}
\alias{qhalfnorm}
\title{Quantile function half Normal}
\usage{
qhalfnorm(p, mean = 0, sd = 1)
}
\description{
Computes the quantiles.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(p, mean = 0, sd = 1) {
  qnorm(
    p * (1 - pnorm(0, mean, sd)) + pnorm(0, mean, sd),
    mean, sd
  )
}
}
\keyword{internal}
