% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_plots.R
\name{qq_plot_censored}
\alias{qq_plot_censored}
\title{Plot the percentile-percentile graph for censored data, using the Turnbull estimator for the empirical cumulative distribution function.}
\usage{
qq_plot_censored(fit)
}
\arguments{
\item{fit}{The result of the fit, obtained through the function MixNRMI1cens or MixNRMI2cens.}
}
\value{
Percentile-percentile graph using the Turnbull estimator
}
\description{
Plot the percentile-percentile graph for censored data, using the Turnbull estimator for the empirical cumulative distribution function.
}
\examples{
set.seed(150520)
data(salinity)
out <- MixNRMI1cens(xleft = salinity$left, xright = salinity$right, extras = TRUE, Nit = 100)
BNPdensity:::qq_plot_censored(out)
}
