% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhalft.R
\name{rhalft}
\alias{rhalft}
\title{Random number generator half Student-t}
\usage{
rhalft(n, df = 1, mean = 0, sd = 1)
}
\description{
Generates a random number.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(n, df = 1, mean = 0, sd = 1) {
  abs(rt_(n, df, mean, sd))
}
}
\keyword{internal}
