% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_.R
\name{rt_}
\alias{rt_}
\title{Random number generator non-standard Student-t}
\usage{
rt_(n, df, mean, sd)
}
\description{
Computes a random number.
}
\details{
For internal use
}
\examples{

## The function is currently defined as
function(n, df, mean, sd) {
  mean + sd * rt(n, df, ncp = 0)
}
}
\keyword{internal}
