% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPdensity-package.R
\docType{data}
\name{salinity}
\alias{salinity}
\title{Salinity tolerance}
\format{
A data frame with 108 observations on the following two variables:
\describe{
\item{left}{A numeric vector.}
\item{right}{A
numeric vector.} }
}
\source{
\code{fitdistrplus} R-package
}
\description{
72-hour acute salinity tolerance (LC50 values) of riverine
macro-invertebrates.
}
\examples{

data(salinity)
hist(salinity$left)
}
\references{
Kefford, B.J., Nugegoda, D., Metzeling, L., Fields, E. 2006.
Validating species sensitivity distributions using salinity tolerance of
riverine macroinvertebrates in the southern Murray-darling Basin (Victoria,
Australia). Canadian Journal of Fisheries and Aquatic Science, 63,
1865-1877.
}
\keyword{datasets}
