% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1.R
\name{summary.NRMI1}
\alias{summary.NRMI1}
\title{S3 method for class 'MixNRMI1'}
\usage{
\method{summary}{NRMI1}(fit, number_of_clusters = FALSE)
}
\arguments{
\item{fit}{A fitted object of class NRMI1}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'MixNRMI1'
}
\examples{
data(acidity)
out <- MixNRMI1(acidity, Nit = 50)
summary(out)
}
