% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI1cens.R
\name{summary.NRMI1cens}
\alias{summary.NRMI1cens}
\title{S3 method for class 'MixNRMI1cens'}
\usage{
\method{summary}{NRMI1cens}(fit, number_of_clusters = FALSE)
}
\arguments{
\item{fit}{A fitted object of class NRMI1cens}
}
\value{
Summary of some relevant information on the fit
}
\description{
S3 method for class 'MixNRMI1cens'
}
\examples{
data(salinity)
out <- MixNRMI1cens(salinity$left, salinity$right, Nit = 50)
summary(out)
}
