% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2.R
\name{summary.NRMI2}
\alias{summary.NRMI2}
\title{S3 method for class 'MixNRMI2'}
\usage{
\method{summary}{NRMI2}(object, number_of_clusters = FALSE, ...)
}
\arguments{
\item{object}{A fitted object of class NRMI2}

\item{number_of_clusters}{Whether to compute the optimal number of clusters, which can be a time-consuming operation (see \code{\link{compute_optimal_clustering}})}

\item{...}{Further arguments to be passed to generic function, ignored at the moment}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'MixNRMI2'
}
\examples{
data(acidity)
out <- MixNRMI2(acidity, Nit = 20)
summary(out)

data(salinity)
out <- MixNRMI2cens(salinity$left, salinity$right, Nit = 20)
summary(out)
}
