% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixNRMI2cens.R
\name{summary.NRMI2cens}
\alias{summary.NRMI2cens}
\title{S3 method for class 'MixNRMI2cens'}
\usage{
\method{summary}{NRMI2cens}(fit, number_of_clusters = FALSE)
}
\arguments{
\item{fit}{A fitted object of class NRMI2cens}
}
\value{
Prints out the text for the summary S3 methods
}
\description{
S3 method for class 'MixNRMI2cens'
}
\examples{
data(salinity)
out <- MixNRMI2cens(salinity$left, salinity$right, Nit = 50)
summary(out)
}
