% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Draw a traceplot for multiple chains}
\usage{
traceplot(fitlist)
}
\arguments{
\item{fitlist}{Output of multMixNRMI.}
}
\value{
A traceplot for multiple chains.
}
\description{
This is a convenience function which works when coda is not yet loaded by the user. If coda is
loaded, it gets masked. See also file multMixNRMI.R
}
