% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Draw a traceplot for multiple chains}
\usage{
traceplot(fitlist)
}
\arguments{
\item{fitlist}{Output of multMixNRMI.}
}
\value{
a coda::mcmc object
}
\description{
Draw a traceplot for multiple chains
}
